Feat: 판매자 확인 및 상품 소유 여부 검증 추가

+++ src/apis/store/product.py
@@ -1,18 +1,31 @@
-from fastapi import Depends, HTTPException
+from fastapi import Cookie, Depends, HTTPException
 
 from src.models.product import Product
-from src.models.repository import ProductRepository
+from src.models.repository import ProductRepository, UserRepository
+from src.models.user import User
 from src.schema.request import CreateProductRequest, UpdateProductRequest
 from src.schema.response import GetProductDetailResponse, GetProductResponse
+from src.service.auth import verify_seller, verify_user_can_access_product
+from src.service.session import SessionService
 
 
 async def create_product_handler(
     request: CreateProductRequest,
     product_repo: ProductRepository = Depends(ProductRepository),
+    user_repo: UserRepository = Depends(UserRepository),
+    session_id: str = Cookie(None),
+    session_service: SessionService = Depends(),
 ) -> GetProductResponse:
-    request_data = request.model_dump(exclude_unset=True)
+    if not session_id:
+        raise HTTPException(status_code=400, detail="Missing Session ID")
 
-    product: Product = Product(**request_data)
+    user_id: int = await verify_seller(
+        session_id=session_id, session_service=session_service
+    )
+    user: User = await user_repo.get_user_by_id(user_id=user_id)
+
+    request_data = request.model_dump(exclude_unset=True)
+    product: Product = Product(seller_id=user.seller.id, **request_data)
     created_product: Product = await product_repo.create_product(product)
 
     return GetProductResponse(
@@ -24,12 +37,28 @@ async def create_product_handler(
 
 
 async def get_product_by_id_handler(
-    product_id: int, product_repo: ProductRepository = Depends(ProductRepository)
+    product_id: int,
+    product_repo: ProductRepository = Depends(ProductRepository),
+    user_repo: UserRepository = Depends(UserRepository),
+    session_id: str = Cookie(None),
+    session_service: SessionService = Depends(),
 ) -> GetProductDetailResponse:
+    if not session_id:
+        raise HTTPException(status_code=400, detail="Missing Session ID")
+
+    user_id: int = await verify_seller(
+        session_id=session_id, session_service=session_service
+    )
+    user: User = await user_repo.get_user_by_id(user_id=user_id)
+
     product: Product | None = await product_repo.get_product_by_id(product_id)
 
     if product is None:
         raise HTTPException(status_code=404, detail="Product Not Found")
+
+    await verify_user_can_access_product(
+        seller_id=user.seller.id, product_seller_id=product.seller_id
+    )
 
     return GetProductDetailResponse(
         id=product.id,
@@ -52,10 +81,25 @@ async def update_product_handler(
     product_id: int,
     request: UpdateProductRequest,
     product_repo: ProductRepository = Depends(ProductRepository),
+    user_repo: UserRepository = Depends(UserRepository),
+    session_id: str = Cookie(None),
+    session_service: SessionService = Depends(),
 ):
+    if not session_id:
+        raise HTTPException(status_code=400, detail="Missing Session ID")
+
+    user_id: int = await verify_seller(
+        session_id=session_id, session_service=session_service
+    )
+    user: User = await user_repo.get_user_by_id(user_id=user_id)
+
     product: Product | None = await product_repo.get_product_by_id(product_id)
 
     if product:
+        await verify_user_can_access_product(
+            seller_id=user.seller.id, product_seller_id=product.seller_id
+        )
+
         request_data = request.model_dump(exclude_unset=True)
         for key, value in request_data.items():
             if hasattr(product, key) and value is not None:
@@ -73,11 +117,27 @@ async def update_product_handler(
 
 
 async def delete_product_handler(
-    product_id: int, product_repo: ProductRepository = Depends(ProductRepository)
+    product_id: int,
+    product_repo: ProductRepository = Depends(ProductRepository),
+    user_repo: UserRepository = Depends(UserRepository),
+    session_id: str = Cookie(None),
+    session_service: SessionService = Depends(),
 ):
+    if not session_id:
+        raise HTTPException(status_code=400, detail="Missing Session ID")
+
+    user_id: int = await verify_seller(
+        session_id=session_id, session_service=session_service
+    )
+    user: User = await user_repo.get_user_by_id(user_id=user_id)
+
     product: Product | None = await product_repo.get_product_by_id(product_id)
 
     if not product:
         raise HTTPException(status_code=404, detail="Product Not Found")
+
+    await verify_user_can_access_product(
+        seller_id=user.seller.id, product_seller_id=product.seller_id
+    )
 
     await product_repo.delete_product(product)
